examples: Renumber and re-document example run scripts
