+++ docs/examples/mfdn_v14/runmfdn.txt
@@ -1,0 +1,79 @@
+
+runmfdn01: harmonic oscillator direct run, i.e., with interaction
+tbmes given directly at hw
+
+    4He Nmax02 hw20 (direct)
+
+    Interaction: run0164-JISP16-tb-10
+
+    Usage: qsubm --here mfdn01
+
+    We use tb-10 as the smallest available set of files.
+
+runmfdn02: same harmonic oscillator run, but with trivial identity
+dilation
+
+    4He Nmax02 hw20 (direct)
+
+    Interaction: run0164-JISP16-tb-10
+
+    Usage: qsubm --here mfdn02
+
+    We use tb-10 as the smallest available set of files, though more
+    commonly files in ob truncation would be used for such a
+    dilation.
+
+runmfdn03: same harmonic oscillator run, but with dilation from
+hw_int=40 (at ob-13 truncation)
+
+    4He Nmax02 hw20  (from hwint=40 at ob-13)
+
+    Interaction: run0164-JISP16-ob-13
+
+    Usage: qsubm --here mfdn03
+
+runmfdn04: harmonic oscillator direct run, at higher Nmax
+
+    4He Nmax12 hw20
+
+    Interaction: run0164-JISP16-tb-20
+
+    Usage (for submission on edison):
+
+        qsubm mfdn04 debug 10 --width=120 --serialthreads=24 --nodesize=24 --opt=--mail-type=ALL
+
+runmfdn05: Nmax and hw scan demo
+
+    4He Nmax-2..20 hw=15..25:5
+
+    Interaction: JISP16/N2LOopt500
+
+    Usage (for submission on edison):
+
+        qsubm mfdn05 debug 10 --width=1 --depth=6 --pool=Nmax02-Mj0.0 --serialthreads=24 --nodesize=24
+        qsubm mfdn05 debug 10 --width=1 --depth=6 --pool=Nmax04-Mj0.0 --serialthreads=24 --nodesize=24
+        qsubm mfdn05 debug 10 --width=1 --depth=6 --pool=Nmax06-Mj0.0 --serialthreads=24 --nodesize=24
+        qsubm mfdn05 debug 10 --width=1 --depth=6 --pool=Nmax08-Mj0.0 --serialthreads=24 --nodesize=24
+        qsubm mfdn05 debug 30 --width=1 --depth=6 --pool=Nmax10-Mj0.0 --serialthreads=24 --nodesize=24
+        qsubm mfdn05 debug 30 --width=6 --depth=6 --pool=Nmax12-Mj0.0 --serialthreads=24 --nodesize=24
+        qsubm mfdn05 debug 30 --width=6 --depth=6 --pool=Nmax14-Mj0.0 --serialthreads=24 --nodesize=24
+        qsubm mfdn05 debug 30 --width=15 --depth=6 --pool=Nmax16-Mj0.0 --serialthreads=24 --nodesize=24
+        qsubm mfdn05 debug 30 --width=66 --depth=6 --pool=Nmax18-Mj0.0 --serialthreads=24 --nodesize=24
+        qsubm mfdn05 debug 30 --width=190 --depth=6 --pool=Nmax20-Mj0.0 --serialthreads=24 --nodesize=24
+
+    Then, to generate archive,
+
+        qsubm mfdn05 --archive
+
+runmfdn06: harmonic oscillator direct run + natural orbitals, with interaction
+tbmes given directly at hw
+
+    4He Nmax02 hw20 (direct)
+    4He Nmax02 hw20 (natorb)
+
+    Interaction: run0164-JISP16-tb-10
+
+    Usage: qsubm --here mfdn06
+
+    We use tb-10 as the smallest available set of files.
+
