+++ docs/examples/runmfdn12.py
@@ -1,19 +1,18 @@
-""" runmfdn12.py
+""" runmfdn13.py
 
     See runmfdn.txt for description.
 
-    Mark A. Caprio
+    Patrick J. Fasano, Mark A. Caprio
     University of Notre Dame
 
-    - 01/04/18 (pjf): Created, copied from runmfd09.
-    - 09/11/19 (pjf):
-        + Remove hard-coded path.
-        + Fix task-data save.
+    - 03/15/19 (pjf): Created, copied from runmfd07.
+    - 09/07/19 (pjf): Remove Nv from truncation_parameters.
 """
 
 import mcscript
 import ncci
 import ncci.mfdn_v15
+import ncci.postprocessing
 
 # initialize mcscript
 mcscript.init()
@@ -23,12 +22,11 @@ mcscript.init()
 ##################################################################
 
 ncci.environ.interaction_run_list = [
-    "run0164-JISP16-ob-9",
-    "run0164-JISP16-ob-13",
-    "run0164-JISP16-tb-10",
-    "run0164-JISP16-tb-20",
-    "run0306-N2LOopt500",  # up to tb-20
-    "runvc0083-Daejeon16-ob-13"
+    "example-data"
+]
+
+ncci.environ.operator_dir_list = [
+    "example-data"
 ]
 
 task = {
@@ -38,13 +36,13 @@ task = {
     # Hamiltonian parameters
     "interaction": "JISP16",
     "use_coulomb": True,
-    "a_cm": 20.,
+    "a_cm": 40.,
     "hw_cm": None,
 
     # input TBME parameters
-    "truncation_int": ("tb", 10),
+    "truncation_int": ("tb", 6),
     "hw_int": 20.,
-    "truncation_coul": ("tb", 10),
+    "truncation_coul": ("tb", 6),
     "hw_coul": 20.,
 
     # basis parameters
@@ -58,19 +56,17 @@ task = {
     "target_truncation": None,
 
     # traditional oscillator many-body truncation
-    "sp_truncation_mode": ncci.modes.SingleParticleTruncationMode.kManual,
-    "mb_truncation_mode": ncci.modes.ManyBodyTruncationMode.kWeightMax,
+    "sp_truncation_mode": ncci.modes.SingleParticleTruncationMode.kNmax,
+    "mb_truncation_mode": ncci.modes.ManyBodyTruncationMode.kNmax,
     "truncation_parameters": {
-        "sp_filename": "${HOME}/tmp/Nmax20-orbitals.dat",
-        "sp_weight_max": 20,
-        "mb_weight_max": 2.1,
-        "parity": +1,
-        "M": 0,
+        "M": 0.0,
+        "Nmax": 6,
+        "Nstep": 2,
         },
 
     # diagonalization parameters
     "diagonalization": True,
-    "eigenvectors": 2,
+    "eigenvectors": 15,
     "initial_vector": -2,
     "max_iterations": 200,
     "tolerance": 1e-6,
@@ -79,12 +75,27 @@ task = {
     # obdme parameters
     ## "hw_for_trans": 20,
     "obdme_multipolarity": 2,
-    "obdme_reference_state_list": [(0, 0, 1)],
+    "obdme_reference_state_list": [(0.0, 0, 1)],
     "save_obdme": True,
+    "ob_observable_sets": ['M1', 'E2'],
 
     # two-body observables
     ## "tb_observable_sets": ["H-components","am-sqr"],
-    "tb_observable_sets": ["H-components"],
+    "tb_observable_sets": ["H-components", "am-sqr"],
+    "tb_observables": [
+        ("CSU3",  (0,0,0), {"CSU3-U": 1/(4-1), "CSU3-V": 1.0}),
+        ("CSp3R", (0,0,0), {"CSp3R-U": 1/(4-1), "CSp3R-V": 1.0}),
+    ],
+    # two-body sources
+    "tbme_sources": [
+        ("CSU3-U", {"filename": "CSU3-U-tb-6.bin", "qn": (0,0,0)}),
+        ("CSU3-V", {"filename": "CSU3-V-tb-6.bin", "qn": (0,0,0)}),
+        ("CSp3R-U", {"filename": "CSp3R-U-tb-6.bin", "qn": (0,0,0)}),
+        ("CSp3R-V", {"filename": "CSp3R-V-tb-6.bin", "qn": (0,0,0)}),
+    ],
+
+    # wavefunction storage
+    "save_wavefunctions": True,
 
     # version parameters
     "h2_format": 15099,
@@ -99,7 +110,7 @@ task = {
 
 # add task descriptor metadata field (needed for filenames)
 task["metadata"] = {
-    "descriptor": ncci.descriptors.task_descriptor_9(task)
+    "descriptor": ncci.descriptors.task_descriptor_7(task)
     }
 
 ncci.radial.set_up_interaction_orbitals(task)
@@ -109,6 +120,8 @@ ncci.radial.set_up_obme_analytic(task)
 ncci.tbme.generate_tbme(task)
 ncci.mfdn_v15.run_mfdn(task)
 ncci.mfdn_v15.save_mfdn_task_data(task)
+ncci.postprocessing.evaluate_ob_observables(task)
+# ncci.handlers.task_handler_oscillator(task)
 
 ##################################################################
 # task control
